Update deleteById-controller
Adding a check for the presence of elements in the board

+++ controllers/boards/deleteById.js
@@ -1,35 +1,36 @@
-const { Board } = require('../../models');
-const { User } = require('../../models');
+const { User, Board } = require('../../models');
 const { HttpError, ctrlWrapper } = require('../../helpers');
 
 const deleteById = async (req, res) => {
   const { boardId } = req.params;
 
-  const result = await Board.findByIdAndDelete(boardId);
+  const board = await Board.findById(boardId);
 
-  if (!result) {
+  if (!board) {
     throw new HttpError(404, 'Not found');
   }
 
-  const userId = req.user._id;
+  if (board.columnOrder.length === 0) {
+    const result = await Board.findByIdAndDelete(boardId);
 
-  const updatedBoards = req.user.boards.filter(
-    board => board.toString() !== boardId
-  );
+    if (!result) {
+      throw new HttpError(404, 'Not found');
+    }
 
-  await User.updateOne({ _id: userId }, { boards: updatedBoards });
+    const userId = req.user._id;
 
-  res.status(200).json({ message: 'Board deleted' });
+    const updatedBoards = req.user.boards.filter(
+      userBoardId => userBoardId.toString() !== boardId
+    );
+
+    await User.updateOne({ _id: userId }, { boards: updatedBoards });
+
+    res.status(200).json({ message: 'Board deleted' });
+  } else {
+    res.status(400).json({ message: 'Cannot delete board with columns' });
+  }
 };
 
 module.exports = {
   deleteById: ctrlWrapper(deleteById),
 };
-
-// if (Array.isArray(columns) && columns.length === 0) {
-//   //   res.status(200).json({ message: 'Board deleted' });
-//   // } else {
-//   //   res
-//   //     .status(400)
-//   //     .json({ message: 'Cannot delete board with columns present' });
-//   // }
